+++ reana-ui/src/components/WorkflowDeleteModal.js
@@ -2,15 +2,14 @@
   -*- coding: utf-8 -*-
 
   This file is part of REANA.
-  Copyright (C) 2020 CERN.
+  Copyright (C) 2020, 2022 CERN.
 
   REANA is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
 */
 
-import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { Button, Modal, Checkbox } from "semantic-ui-react";
+import { Button, Modal, Message, Icon } from "semantic-ui-react";
 
 import { deleteWorkflow, closeDeleteWorkflowModal } from "~/actions";
 import {
@@ -20,7 +19,6 @@ import {
 
 export default function WorkflowDeleteModal() {
   const dispatch = useDispatch();
-  const [deleteWorkspace, setDeleteWorkspace] = useState(true);
   const open = useSelector(getWorkflowDeleteModalOpen);
   const workflow = useSelector(getWorkflowDeleteModalItem);
 
@@ -28,32 +26,32 @@ export default function WorkflowDeleteModal() {
 
   const onCloseModal = () => {
     dispatch(closeDeleteWorkflowModal());
-    setDeleteWorkspace(true);
   };
 
   const { id, name, run, size } = workflow;
   return (
-    <Modal open={open} onClose={onCloseModal} closeIcon>
+    <Modal open={open} onClose={onCloseModal} closeIcon size="small">
       <Modal.Header>Delete workflow</Modal.Header>
       <Modal.Content>
-        <>
-          <p>
-            Are you sure you want to delete workflow "{name} #{run}"?
-          </p>
-          <Checkbox
-            checked={deleteWorkspace}
-            onChange={(_, data) => setDeleteWorkspace(data.checked)}
-            label={`Delete also workflow workspace ${
-              size.human_readable ? `(free up ${size.human_readable})` : ""
-            } `}
-          />
-        </>
+        <Message icon warning>
+          <Icon name="warning sign" />
+          <Message.Content>
+            <Message.Header>Workspace deletion!</Message.Header>
+            This action will delete also the workflow's workspace
+            {size.human_readable ? ` (${size.human_readable})` : ""}. Please
+            make sure to download all the files you want to keep before
+            proceeding.
+          </Message.Content>
+        </Message>
+        <p>
+          Are you sure you want to delete workflow "{name} #{run}"?
+        </p>
       </Modal.Content>
       <Modal.Actions>
         <Button
           negative
           onClick={() => {
-            dispatch(deleteWorkflow(id, deleteWorkspace)).then(() => {
+            dispatch(deleteWorkflow(id)).then(() => {
               onCloseModal();
             });
           }}